Convert project to typescript
